requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==8.0.1
django-colorful==1.3
django-js-asset==2.0.0
djangorestframework==3.13.1
django-ckeditor==5.8.0